standalone/requirements.txt
# Frigate Notify Bridge - Standalone Requirements

# Web server
aiohttp>=3.9.0
uvloop>=0.19.0;platform_system!="Windows"

# MQTT
paho-mqtt>=2.0.0

# Push providers
firebase-admin>=6.0.0

# QR code generation
qrcode>=7.4.0
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
